refactor to add FormWarehouse component
