+++ src/components/FormWarehouse/FormWarehouse.scss
@@ -1,0 +1,108 @@
+@use "../../styles/partials/variables" as *;
+@use "../../styles/partials/mixins" as *;
+
+.form-wrapper {
+  & * {
+    margin: 0;
+  }
+}
+
+.form-fields {
+  @include tablet {
+    display: flex;
+    padding: 1.5rem 2rem 2rem;
+  }
+
+  @include laptop {
+    padding: 1.5rem 2.5rem 2rem;
+  }
+}
+
+.form-section {
+  padding: 1rem 1.5rem 1.5rem;
+
+  @include tablet {
+    width: 50%;
+    padding: 0;
+    padding-left: 2rem;
+  }
+
+  @include laptop {
+    padding-left: 2.5rem;
+  }
+
+  &--warehouse {
+    border-bottom: 1px solid $cloud;
+
+    @include tablet {
+      border-bottom: none;
+      border-right: 1px solid $cloud;
+      padding-right: 2rem;
+      padding-left: 0;
+    }
+
+    @include laptop {
+      padding-right: 2.5rem;
+      padding-left: 0;
+    }
+  }
+
+  &__input-wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+    @include tablet {
+      gap: 1.5rem;
+    }
+  }
+
+  &__header {
+    margin-bottom: 1.5rem;
+    
+    @include tablet {
+      margin-bottom: 2rem;      
+    }
+  }
+}
+
+.form-label {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.form-input {
+  @include interactive-components;
+  border: 1px solid $cloud;
+
+  &:focus {
+    border: 1px solid $instock-indigo;
+    outline: none;
+  }
+
+  &--error {
+    @extend .form-input;
+    border: 1px solid $red;
+  }
+}
+
+.buttons {
+  padding: 1rem 1.5rem;
+  background-color: $light-grey;
+  display: flex;
+  gap: 1rem;
+
+  @include tablet {
+    justify-content: flex-end;
+    padding: 1rem 2.5rem;
+  }
+
+  .btn__add--form {
+    width: 50%;
+
+    @include tablet {
+      width: 8.625rem;
+    }
+  }
+}
